Revert "Add automatic remote build detection for Python apps in func pack command (#4530)"
This reverts commit b791c6ebd6de5511984f018b2351b585d2a935b3.
