+++ src/Cli/func/Helpers/PublishHelper.cs
@@ -1,7 +1,8 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.Net.Http.Handlers;
+using Azure.Functions.Cli.Arm.Models;
 using Azure.Functions.Cli.Common;
 
 namespace Azure.Functions.Cli.Helpers
@@ -23,6 +24,33 @@ namespace Azure.Functions.Cli.Helpers
             }
 
             return null;
+        }
+
+        public static BuildOption ResolveBuildOption(BuildOption currentBuildOption, WorkerRuntime runtime, Site site, bool buildNativeDeps, bool noBuild)
+        {
+            // --no-build and --build-native-deps will take precedence over --build local and --build remote
+            if (noBuild)
+            {
+                return BuildOption.None;
+            }
+
+            if (buildNativeDeps)
+            {
+                return BuildOption.Container;
+            }
+
+            if (currentBuildOption == BuildOption.Default)
+            {
+                // Change to remote build if, python app, has requirements.txt, requirements.txt has content
+                if (runtime == WorkerRuntime.Python &&
+                    FileSystemHelpers.FileExists(Constants.RequirementsTxt) &&
+                    new FileInfo(Path.Combine(Environment.CurrentDirectory, Constants.RequirementsTxt)).Length > 0)
+                {
+                    return BuildOption.Remote;
+                }
+            }
+
+            return currentBuildOption;
         }
 
         public static async Task<HttpResponseMessage> InvokeLongRunningRequest(HttpClient client, ProgressMessageHandler handler, HttpRequestMessage request, long requestSize = 0, string prompt = null)
